Updated fieldNumber initialization and increment logic
The initial value of the variable 'fieldNumber' has been changed from 0 to 1. Also, the increment operation for 'fieldNumber' is now done separately after adding a field to 'metaType', instead of being part of the AddField method call. This change ensures that the correct field number is used when adding fields to 'metaType'.